move files and test path
