+++ earthnet_models_pytorch/data/de23_data.py
@@ -18,93 +18,63 @@ from torch.utils.data import Dataset, DataLoader, random_split
 from earthnet_models_pytorch.utils import str2bool
 
 variables = {
+    ### variables with dimensions x, y, Sampled time (irregular)
     # Sentinel 2 bands. Spatio-temporal data. Every 5 days.
-    "s2_bands": ["s2_B02", "s2_B03", "s2_B04", "s2_B05", "s2_B06", "s2_B07", "s2_B8A"],
-    "s2_avail": ["s2_avail"],
-    "s2_scene_classification": ["s2_SCL"],
-    "cloud_mask": ["s2_mask"],
-    # Sentinel 1 bands. Spatio-temporal data. Every 12 days.
-    "s1_bands": ["s1_vv", "s1_vh"],
-    "s1_avail": ["s1_avail"],
-    # Monthly mean and std computed over 1984-2020 using Landsat 5, 7, 8. Digital Earth Africa product.
-    "ndviclim": ["ndviclim_mean", "ndviclim_std"],
-    # Era5 reanalysis dataset. Era5 era5land pev and pet have a low precision
-    "era5lands": [
-        "era5land_t2m_mean",
-        "era5land_pev_mean",
-        "era5land_slhf_mean",
-        "era5land_ssr_mean",
-        "era5land_sp_mean",
-        "era5land_sshf_mean",
-        "era5land_e_mean",
-        "era5land_tp_mean",
-        "era5land_t2m_min",
-        "era5land_pev_min",
-        "era5land_slhf_min",
-        "era5land_ssr_min",
-        "era5land_sp_min",
-        "era5land_sshf_min",
-        "era5land_e_min",
-        "era5land_tp_min",
-        "era5land_t2m_max",
-        "era5land_pev_max",
-        "era5land_slhf_max",
-        "era5land_ssr_max",
-        "era5land_sp_max",
-        "era5land_sshf_max",
-        "era5land_e_max",
-        "era5land_tp_max",
-    ],
-    # Era5 recomputed by Fabian Gans for the XAIDA project. More accurate.
+    "s2_bands": ["B02", "B03", "B04", "B05", "B06", "B07", "B8A"],
+    # "s2_avail": ["s2_avail"],
+    "s2_scene_classification": ["SCL"],
+    # S2 EarthNet cloudmask
+    "cloud_mask": ["cloudmask_en"],
+    ### variables with dimension time
+    # Era5 reanalysis dataset. Era5 pev has a low precision
     "era5": [
-        "era5_e",
-        "era5_pet",
-        "era5_pev",
-        "era5_ssrd",
-        "era5_t2m",
-        "era5_t2mmax",
-        "era5_t2mmin",
-        "era5_tp",
+        "t2m_mean",
+        "pev_mean",
+        "slhf_mean",
+        "ssr_mean",
+        "sp_mean",
+        "sshf_mean",
+        "e_mean",
+        "tp_mean",
+        "t2m_min",
+        "pev_min",
+        "slhf_min",
+        "ssr_min",
+        "sp_min",
+        "sshf_min",
+        "e_min",
+        "tp_min",
+        "t2m_max",
+        "pev_max",
+        "slhf_max",
+        "ssr_max",
+        "sp_max",
+        "sshf_max",
+        "e_max",
+        "tp_max",
     ],
-    # soigrids is not yet define on 128 x 128 pixels. Can maybe not be used during the prediction period.
-    "soilgrids": [
-        "sg_bdod_top_mean",
-        "sg_bdod_sub_mean",
-        "sg_cec_top_mean",
-        "sg_cec_sub_mean",
-        "sg_cfvo_top_mean",
-        "sg_cfvo_sub_mean",
-        "sg_clay_top_mean",
-        "sg_clay_sub_mean",
-        "sg_nitrogen_top_mean",
-        "sg_nitrogen_sub_mean",
-        "sg_phh2o_top_mean",
-        "sg_phh2o_sub_mean",
-        "sg_ocd_top_mean",
-        "sg_ocd_sub_mean",
-        "sg_sand_top_mean",
-        "sg_sand_sub_mean",
-        "sg_silt_top_mean",
-        "sg_silt_sub_mean",
-        "sg_soc_top_mean",
-        "sg_soc_sub_mean",
+    ### variables with dimension event_time
+    "event" : [
+        "events",
+        "event_labels",
     ],
-    # Elevation map. Defined on 0 - 2000.
-    "elevation": ["cop_dem"],  # ['srtm_dem', 'alos_dem', 'cop_dem'],
-    # Landscape classes. Categorical variable.
-    "landcover": ["esawc_lc"],
-    # Geomorphon classes. Categorical variable. geomorphon is not define on 128 x 128 pixels neither.
-    "geomorphon": ["geom_cls"],
+    ### static variables
+    ### variables with dimensions x_300, y_300, 
+    # Land cover classes. Categorical variable.
+    "landcover": ["lccs_class"],
+    ### variables with dimensions x, y
+    # Elevation model. Defined on 0 - 2000.
+    "elevation": ["cop_dem"], 
 }
 
 
-class EarthNet2023Dataset(Dataset):
+class DeepExtremes2023Dataset(Dataset):
     def __init__(
         self, folder: Union[Path, str], target: str, variables=variables, fp16=False
     ):
         if not isinstance(folder, Path):
             folder = Path(folder)
-        self.filepaths = sorted(list(folder.glob("*/*.nc")))  # why sorted?
+        self.filepaths = sorted(list(folder.glob("*.zarr")))  # why sorted?
         self.type = np.float16 if fp16 else np.float32
         self.target = target
         self.variables = variables
@@ -115,31 +85,32 @@ class EarthNet2023Dataset(Dataset):
         # print(filepath)
         minicube = xr.open_dataset(filepath)
 
-        s2_avail = np.squeeze(minicube[self.variables["s2_avail"]].to_array(), axis=0)
-
-        # s2 is 5 days is every 5 days
-        t_len = s2_avail.shape[0]
-        time = [i for i in range(4, t_len, 5)]
-
-        index_avail = np.where(s2_avail.values.astype(self.type) == 1)[0]
-        if index_avail[0] % 5 != 4:
+        print(minicube.cloud_mask.time)
+        print(minicube.s2_bands.B02.time)
+        if minicube.cloud_mask.time != minicube.s2_bands.B02.time:
             raise Exception(
                 "The first available imagery of sentinel-2 is not 4 + [5]"
                 + str(index_avail)
             )
 
-        # Select the days with available data
-        s2_avail = s2_avail.isel(time=time).values  # .astype(int)
 
         # Create the minicube
         # s2 is 5 days, and already rescaled [0, 1]
+        print(minicube[self.variables["s2_bands"]].shape)
         s2_cube = (
             minicube[self.variables["s2_bands"]]
             .to_array()
-            .isel(time=time)
-            .values.transpose((1, 0, 2, 3))
-            .astype(self.type)[:, :, :128, :128]
+            .values.transpose((3, 0, 1, 2))
+            .astype(self.type)
         )  # (time, channels, w, h)
+        print(s2_cube)
+
+        # s2_mask: 
+        # 0 - free_sky
+        # 1 - cloud
+        # 2 - cloud_shadows
+        # 3 - snow
+        # 4 - masked_other_reasons
 
         s2_mask = (
             minicube[self.variables["cloud_mask"]]
@@ -159,17 +130,18 @@ class EarthNet2023Dataset(Dataset):
         # weather is daily
         meteo_cube = minicube[self.variables["era5"]]
 
-        # rescale temperature on the extreme values ever observed in Africa (Kelvin).
-        meteo_cube["era5_t2m"] = (meteo_cube["era5_t2m"] - 248) / (328 - 248)
-        meteo_cube["era5_t2mmin"] = (meteo_cube["era5_t2mmin"] - 248) / (328 - 248)
-        meteo_cube["era5_t2mmax"] = (meteo_cube["era5_t2mmax"] - 248) / (328 - 248)
+        # rescale temperature on the extreme values ever observed globally (Kelvin): -88, 58.
+        meteo_cube["t2m"] = (meteo_cube["t2m"] - 185) / (331 - 185)
+        meteo_cube["t2mmin"] = (meteo_cube["t2mmin"] - 185) / (331 - 185)
+        meteo_cube["t2mmax"] = (meteo_cube["t2mmax"] - 185) / (331 - 185)
 
+        # shouldn't I rescale all meteo variables?
         meteo_cube = (
             meteo_cube.to_array().values.transpose((1, 0)).astype(self.type)
         )  # t, c
         meteo_cube[np.isnan(meteo_cube)] = 0
 
-        # TODO NaN values are replaces by the mean of each variable. To solve, currently RuntimeWarning: overflow encountered in reduce
+        # TODO NaN values are replaced by the mean of each variable. To solve, currently RuntimeWarning: overflow encountered in reduce
         # col_mean = np.nanmean(meteo_cube, axis=0)
         # inds = np.where(np.isnan(meteo_cube))
         # meteo_cube[inds] = np.take(col_mean, inds[1])
@@ -193,13 +165,25 @@ class EarthNet2023Dataset(Dataset):
         # NaN values handling
         s2_cube = np.where(np.isnan(s2_cube), np.zeros(1).astype(self.type), s2_cube)
         target = np.where(np.isnan(target), np.zeros(1).astype(self.type), target)
+        s2_mask = np.where(np.isnan(s2_mask), np.ones(1).astype(self.type), s2_mask) # ?? s2_cube ? or s2_mask ?
+
+        landcover = np.where(
+            np.isnan(landcover), np.zeros(1).astype(self.type), landcover
+        )
+
+        # include scene classification in model? https://sentinels.copernicus.eu/web/sentinel/technical-guides/sentinel-2-msi/level-2a/algorithm-overview
+        # Concatenation
+        satellite_data = np.concatenate((target, s2_cube), axis=1)
 
         # Final minicube
         data = {
-            "dynamic": [torch.from_numpy(s2_cube), torch.from_numpy(meteo_cube)],
+            "dynamic": [
+                torch.from_numpy(satellite_data),
+                torch.from_numpy(meteo_cube),
+            ],
             "dynamic_mask": [torch.from_numpy(s2_mask)],
             "static": [torch.from_numpy(topography)],
-            "target": torch.from_numpy(target),
+            # "target": torch.from_numpy(target),
             # "s2_avail": torch.from_numpy(s2_avail),
             "landcover": torch.from_numpy(landcover),
             "filepath": str(filepath),
@@ -231,8 +215,8 @@ class EarthNet2023Dataset(Dataset):
     def target_computation(self, minicube):
         """Compute the vegetation index (VI) target"""
         if self.target == "ndvi":
-            targ = (minicube.s2_B8A - minicube.s2_B04) / (
-                minicube.s2_B8A + minicube.s2_B04 + 1e-6
+            targ = (minicube.B8A - minicube.B04) / (
+                minicube.B8A + minicube.B04 + 1e-6
             )
 
         if (
@@ -240,17 +224,29 @@ class EarthNet2023Dataset(Dataset):
         ):  # TODO the the denominator is not optimal, needs to be improved accordingly to the original paper
             targ = np.tanh(
                 (
-                    (minicube.s2_B08 - minicube.s2_B04)
-                    / (minicube.s2_BO8 + minicube.s2_B04 + 1e-6)
+                    (minicube.B08 - minicube.B04)
+                    / (minicube.BO8 + minicube.B04 + 1e-6)
                 )
                 ** 2
             ) / np.tanh(1)
 
-        # TODO add ndvi normalized, for each measurement, need to select the good month in ndviclim
+        # if self.target == "anomalie_ndvi":
+        #     targ = (minicube.B8A - minicube.B04) / (
+        #         minicube.B8A + minicube.B04 + 1e-6
+        #     )
+        #     for i in range(1, 13):
+        #         indices = targ.groupby("time.month").groups[i]
+        #         index_month = minicube.ndviclim_mean.groupby("time_clim.month").groups[
+        #             i
+        #         ]
+        #         targ[indices] = (
+        #             targ[indices].values - minicube.ndviclim_mean[index_month].values
+        #         )
+
         return targ
 
 
-class EarthNet2023DataModule(pl.LightningDataModule):
+class DeepExtremes2023DataModule(pl.LightningDataModule):
     def __init__(self, hparams: argparse.Namespace):
         super().__init__()
         self.save_hyperparameters(copy.deepcopy(hparams))
@@ -287,8 +283,8 @@ class EarthNet2023DataModule(pl.LightningDataModule):
 
     def setup(self, stage: str = None):
         if stage == "fit" or stage is None:
-            earthnet_full = EarthNet2023Dataset(
-                self.base_dir / "train",
+            earthnet_full = DeepExtremes2023Dataset(
+                self.base_dir, # / "train",
                 target=self.hparams.target,
                 fp16=self.hparams.fp16,
             )
@@ -299,7 +295,7 @@ class EarthNet2023DataModule(pl.LightningDataModule):
             )
 
         if stage == "test" or stage is None:
-            self.earthnet_test = EarthNet2023Dataset(
+            self.earthnet_test = DeepExtremes2023Dataset(
                 self.base_dir / "test",
                 target=self.hparams.target,
                 fp16=self.hparams.fp16,
